requirements.txt
pandas>=2.2.0
selenium>=4.21.0
tqdm>=4.66.0
webdriver-manager>=4.0.1
Unidecode>=1.3.8
google-generativeai>=0.7.2
python-dotenv>=1.0.1
